Features -------- - Added support for --use-unknown to map all n-grams containing OOV words to <unk>. - Unigrams backoff to uniform distribution across all vocabulary (including <unk>). - Verified --read-vocab filters the text/count input to only those n-grams containing non-OOV words.
Cleanup
-------
- Replaced BeginningOfSentence with EndOfSentence to reduce special cases.

Notes
-----
- The --read-vocab filter behaves differently from SRILM.  SRILM appears to filter
out n-grams with OOV as target word initially and later removes ones with OOV in the
n-gram history when writing out the LM.  Thus, the count statistics used to estimate
the probabilities are slightly different.

To Do
-----
- Need to reintroduce test cases with a small data set to make sure nothing is
broken.  The code base at this point is not sufficiently tested.




git-svn-id: http://mitlm.googlecode.com/svn/trunk@20 4622ca98-d053-0410-b680-d1345939a8d7